+++ sdk/client.py
@@ -13,127 +13,22 @@
 # limitations under the License.
 from __future__ import annotations
 
-from abc import abstractmethod
-from datetime import datetime, timedelta, timezone
+from getpass import getpass
 from typing import Any, Dict, List, Optional, Tuple
 
-import requests
-from jwt import decode, DecodeError
-from requests import PreparedRequest
 from requests.auth import AuthBase
+import requests
 
-from sdk.endpoints import Endpoints
+from sdk.authentication import (
+    TokenProvider,
+    Auth0TokenProvider,
+    HTTPBearerAuthenticator,
+)
+from sdk.endpoints import Endpoints, Auth0Conf
 from sdk.errors import HTTPError
 from sdk.utils.jsend import JSendPayload
 
 TIMEOUT = 30  # client http requests timeout after 30s
-
-
-class HTTPBearerAuthenticator(AuthBase):
-    def __init__(self, token_provider: Optional[TokenProvider]):
-        if not token_provider:
-            raise Exception("The authenticator needs a token provider.")
-        self.token_provider = token_provider
-
-    def __call__(self, r: PreparedRequest) -> PreparedRequest:
-        r.headers["Authorization"] = f"Bearer {self.token_provider.get_token()}"
-        return r
-
-
-class TokenProviderError(Exception):
-    pass
-
-
-class TokenProvider:
-    __token_cache: Optional[tuple[datetime, str]] = None
-    expiry_window: timedelta = timedelta(minutes=1.0)
-
-    @abstractmethod
-    def _query_token(self) -> dict[str, Any]:
-        raise NotImplementedError
-
-    def get_token(self) -> str:
-        if self.__token_cache:
-            expiry, token = self.__token_cache
-            if expiry - self.expiry_window > datetime.now(tz=timezone.utc):
-                return token
-
-        self.__token_cache = self._refresh_token_cache()
-        return self.__token_cache[1]
-
-    def _refresh_token_cache(self) -> tuple[datetime, str]:
-        try:
-            token_response = self._query_token()
-            expiry = self._extract_expiry(token_response)
-
-            if expiry is None:
-                expiry = datetime.now(tz=timezone.utc)
-
-            return (expiry, token_response["access_token"])
-        except HTTPError as err:
-            raise TokenProviderError(err)
-
-    @staticmethod
-    def _extract_expiry(token_response: dict[str, Any]) -> datetime | None:
-        """Extracts the expiry datetime from the token response.
-
-        Assumes that the actual token provided is correct, but might be in an
-        unexpected format
-
-        Args:
-            token_response: A valid token response dictionary.
-
-        Returns:
-            The time the token will expire, or None if it can't be calculated
-        """
-        expires_in = token_response.get("expires_in", None)
-        if expires_in:
-            return datetime.now(tz=timezone.utc) + timedelta(seconds=float(expires_in))
-
-        try:
-            # We assume the token is valid, but might not be in an expected format
-            token = token_response.get("access_token")
-
-            if isinstance(token, str) and "." in token:
-                token_timestamp = decode(
-                    token,
-                    algorithms=["HS256"],
-                    options={"verify_signature": False},
-                ).get("exp")
-                if token_timestamp:
-                    return datetime.fromtimestamp(token_timestamp, tz=timezone.utc)
-
-        except DecodeError:
-            # The token is not in a format that could be parsed as a JWT
-            pass
-        return None
-
-
-class UsernamePasswordTokenProvider(TokenProvider):
-    def __init__(self, username: str, password: str, login_url: str):
-        self.username = username
-        self.password = password
-        self.login_url = login_url
-
-    def _query_token(self) -> Dict[str, Any]:
-        payload = {
-            "email": self.username,
-            "password": self.password,
-            "type": "user",
-        }
-
-        rsp = requests.post(
-            self.login_url,
-            json=payload,
-            timeout=TIMEOUT,
-            headers={"content-type": "application/json"},
-        )
-        data = rsp.json()["data"]
-
-        if rsp.status_code >= 400:
-            raise HTTPError(data)
-
-        return {"access_token": data["token"]}
 
 
 class Client:
@@ -146,21 +41,65 @@ class Client:
         password: Optional[str] = None,
         token_provider: Optional[TokenProvider] = None,
         endpoints: Optional[Endpoints] = None,
+        auth0: Optional[Auth0Conf] = None,
     ):
-        if not (username and password) and not token_provider:
-            raise Exception(
-                "At least a username/password combination or"
-                "TokenProvider object should be provided."
+        if not username and not token_provider:
+            raise ValueError(
+                "At least a username or TokenProvider object should be provided."
             )
-        self.endpoints = endpoints or Endpoints()
-        self.group_id = group_id
+        if token_provider is not None:
+            self._check_token_provider(token_provider)
 
-        if username and password:
-            token_provider = UsernamePasswordTokenProvider(
-                username, password, f"{self.endpoints.account}/api/v1/auth/login"
-            )
+        self.endpoints = self._make_endpoints(endpoints)
+
+        if username:
+            auth0 = self._make_auth0(auth0)
+            token_provider = self._credential_login(username, password, auth0)
 
         self.authenticator = HTTPBearerAuthenticator(token_provider)
+        self.group_id = group_id
+
+    @staticmethod
+    def _make_endpoints(endpoints: Optional[Endpoints]) -> Endpoints:
+        if endpoints is None:
+            return Endpoints()
+
+        if not isinstance(endpoints, Endpoints):
+            raise TypeError(f"endpoints must be a {Endpoints.__name__} instance")
+
+        return endpoints
+
+    @staticmethod
+    def _make_auth0(auth0: Optional[Auth0Conf]) -> Auth0Conf:
+        if auth0 is None:
+            return Auth0Conf()
+
+        if not isinstance(auth0, Auth0Conf):
+            raise TypeError(f"auth0 parameter must be a {Auth0Conf.__name__} instance")
+
+        return auth0
+
+    @staticmethod
+    def _check_token_provider(token_provider: TokenProvider) -> None:
+        try:
+            # The type ignore is because I wouldn't know how to fix the type problem
+            # but the code should be correct, and is tested
+            issubclass(token_provider, TokenProvider)  # type: ignore
+        except TypeError:
+            raise TypeError("token_provider must be a TokenProvider subclass")
+
+    def _credential_login(
+        self, username: str, password: Optional[str], auth0: Auth0Conf
+    ) -> TokenProvider:
+        if not password:
+            password = getpass("Enter your password:")
+            #   We want to allow an empty string at first, but then
+            #   it results in an error
+            if not password:
+                raise ValueError("The prompted password should not be empty")
+
+        token_provider: TokenProvider = Auth0TokenProvider(username, password, auth0)
+        return token_provider
 
     def _request(
         self, method: str, url: str, payload: Optional[Dict[str, Any]] = None