+++ tests/test_device_specs.py
@@ -1,13 +1,16 @@
 import json
 from uuid import uuid4
+from unittest.mock import patch
 
 import pytest
 
 from sdk import SDK
+from tests.test_doubles.authentication import FakeAuth0AuthenticationSuccess
 
 
 class TestDeviceSpecs:
     @pytest.fixture(autouse=True)
+    @patch("sdk.client.Auth0TokenProvider", FakeAuth0AuthenticationSuccess)
     def init_sdk(self, start_mock_request):
         self.sdk = SDK(
             username="me@test.com", password="password", group_id=str(uuid4())