+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.15] 2023-03-27
+
+### Added
+    - Added tests to check the login behavior.
+    - Added tests to check the override Endpoints behavior.
+
+### Changed
+    - The authentication now directly connects to the Auth0 platform instead of connecting through PasqalCloud.
+    - Small refactor of files, with the authentication modules in the `authentication.py` file, instead of `client.py`.
+
+### Deleted
+    - Account endpoint, we now use Auth0.
+
 ## [0.1.14]
 
 ### Changed